Move socket handling into its own file
